Refactored my code and resolved all comments on my PR
