+++ src/DwitTech.AccountService.Core/Interfaces/IAuthenticationService.cs
@@ -7,5 +7,7 @@ namespace DwitTech.AccountService.Core.Interfaces
     {
         Task<TokenModel> GenerateAccessToken(User user);
         Task<TokenModel> GenerateAccessTokenFromRefreshToken(TokenModel tokenModel);
+        Task<TokenModel> AuthenticateUserLogin(string email, string hashedPassword);
+
     }
 }
