+++ src/DwitTech.AccountService.Core/Interfaces/IUserService.cs
@@ -1,7 +1,3 @@
-using DwitTech.AccountService.Core.Dtos;
-using DwitTech.AccountService.Core.Models;
-using DwitTech.AccountService.Data.Entities;
-using Microsoft.AspNetCore.Http;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -12,7 +8,6 @@ namespace DwitTech.AccountService.Core.Interfaces
 {
     public interface IUserService
     {
-        Task<TokenModel> AuthenticateUserLogin(string email, string hashedPassword);
     }
 }
