+++ src/DwitTech.AccountService.Data/Context/AccountDbContext.cs
@@ -24,7 +24,7 @@ namespace DwitTech.AccountService.Data.Context
         public DbSet<Role> Roles { get; set; }
         public DbSet<SessionToken> SessionTokens { get; set; }
 
-        public DbSet<UserLogin> UsersLogin { get; set; }
+        public DbSet<UserLogin> UserLogins { get; set; }
 
         protected override void OnModelCreating(ModelBuilder builder)
         {
