+++ src/DwitTech.AccountService.Data/repository/IUserRepository.cs
@@ -8,10 +8,7 @@ namespace DwitTech.AccountService.Data.Repository
         Task SaveUserValidationCode(ValidationCode validation);
         Task<ValidationCode> GetUserValidationCode(string activationCode, CodeType codeType);
         Task<User> GetUser(int id);
-        Task<User> GetUserByEmail(string email);
         Task UpdateUser(User user);
-        Task<bool> ValidateLogin(string email, string hashedPassword);
-        
     }
 
 }
