+++ tests/DwitTech.AccountService.WebApi.Tests/Controllers/UserControllerTest.cs
@@ -13,9 +13,6 @@ namespace DwitTech.AccountService.WebApi.Tests.Controllers
 {
     public class UserControllerTest
     {
-        private readonly Mock<IActivationService> _mockActService;
-        private readonly UserController _controller;
-        private readonly Mock<IUserService> _userService;
         private readonly IConfiguration _configuration;
 
         public UserControllerTest()
@@ -23,9 +20,7 @@ namespace DwitTech.AccountService.WebApi.Tests.Controllers
             _configuration = new ConfigurationBuilder().AddInMemoryCollection(new Dictionary<string, string>
             {
                 { "FROM_EMAIL","example@gmail.com" },
-                { "NOTIFICATION_SERVICE_SENDMAIL_END_POINT", "https://jsonplaceholder.typicode.com/posts"},
-                {"X_API_KEY", "your_api_key"},
-                {"SOURCE_IP", "127.0.0.1"}
+                { "NOTIFICATION_SERVICE_SENDMAIL_END_POINT", "https://jsonplaceholder.typicode.com/posts"}
 
             }).Build();
         }
@@ -40,13 +35,10 @@ namespace DwitTech.AccountService.WebApi.Tests.Controllers
             var mockDbContext = new Mock<AccountDbContext>(options);
             var userRepository = new Mock<UserRepository>(mockDbContext.Object);
             var iHttpClientFactory = new Mock<IHttpClientFactory>();
+            var authRepository = new Mock<AuthenticationRepository>(mockDbContext.Object);
             var _mockService = new ActivationService(_configuration, userRepository.Object, iHttpClientFactory.Object);
-            var authenticationService = new Mock<IAuthenticationService>();
-            var mockNext = new Mock<RequestDelegate>();
-            var authorizationMiddleware = new Mock<AuthorizationMiddleware>(mockNext.Object, _configuration);
-            var httpContextAccessor = new Mock<IHttpContextAccessor>();
-            var _userService = new Mock<UserService>(userRepository.Object, _configuration, authenticationService.Object, authorizationMiddleware.Object, httpContextAccessor.Object);
-            var userController = new UserController(_mockService, _userService.Object);
+            var _mockAuthService = new AuthenticationService(_configuration, authRepository.Object);
+            var userController = new UserController(_mockService, _mockAuthService);
             string activationCode = "erg3345dh2";
 
             //act
@@ -66,14 +58,11 @@ namespace DwitTech.AccountService.WebApi.Tests.Controllers
             var mockDbContext = new Mock<AccountDbContext>(options);
             var userRepository = new Mock<UserRepository>(mockDbContext.Object);
             var iHttpClientFactory = new Mock<IHttpClientFactory>();
-            var authenticationService = new Mock<IAuthenticationService>();
-            var mockNext = new Mock<RequestDelegate>();
-            var authorizationMiddleware = new Mock<AuthorizationMiddleware>(mockNext.Object, _configuration);
-            var httpContextAccessor = new Mock<IHttpContextAccessor>();
+            var authRepository = new Mock<AuthenticationRepository>(mockDbContext.Object);
             var _mockService = new Mock<ActivationService>(_configuration, userRepository.Object, iHttpClientFactory.Object);
-            var _userService = new Mock<UserService>(userRepository.Object, _configuration, authenticationService.Object, authorizationMiddleware.Object, httpContextAccessor.Object);
+            var _mockAuthService = new AuthenticationService(_configuration, authRepository.Object);
 
-            var userController = new UserController(_mockService.Object, _userService.Object);
+            var userController = new UserController(_mockService.Object, _mockAuthService);
 
             string email = "hello@support.com";
             string hashedPassword = "whgwygy37t63t36shhcxvw";
